+++ foodbook_app/lib/presentation/widgets/restaurant_reviews/review_item.dart
@@ -45,25 +45,25 @@ class ReviewItem extends StatelessWidget {
             Row(
               children: [
                 const Expanded(child: Text('Cleanliness')),
-                _buildRatingStars(review.ratings['cleanliness'] ?? 0),
+                _buildRatingStars((review.ratings['cleanliness'] ?? 0).toInt()),
               ],
             ),
             Row(
               children: [
                 const Expanded(child: Text('Waiting Time')),
-                _buildRatingStars(review.ratings['waitingTime'] ?? 0),
+                _buildRatingStars((review.ratings['waitingTime'] ?? 0).toInt()),
               ],
             ),
             Row(
               children: [
                 const Expanded(child: Text('Service')),
-                _buildRatingStars(review.ratings['service'] ?? 0),
+                _buildRatingStars((review.ratings['service'] ?? 0).toInt()),
               ],
             ),
             Row(
               children: [
                 const Expanded(child: Text('Food Quality')),
-                _buildRatingStars(review.ratings['foodQuality'] ?? 0),
+                _buildRatingStars((review.ratings['foodQuality'] ?? 0).toInt()),
               ],
             ),
             const SizedBox(height: 10),
